Приступил к разделению ответственности
- Вынес моки в отдельный файл(класс) QuestionFactory.swift
- Создал этому файлу интерфейс взаимодействия в виде функции requestNextQuestion ()
- Создал раздел проекта Models и вынес в него:
- QuizStepViewModel
- QuizResultsViewModel
- QuizQuestion
в виде отдельных файлов.
