+++ Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Nikita Belov on 28.01.23.
+//
+
+import Foundation
+
+/// Вью-модель для состояния "Результат квиза"
+private struct QuizResultsViewModel {
+	let title: String
+	let text: String
+	let buttonText: String
+}
