+++ Models/QuizStepViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Nikita Belov on 28.01.23.
+//
+
+import UIKit
+
+/// Вью-модель для состояния "Вопрос задан"
+private struct QuizStepViewModel {
+	let image: UIImage
+	let question: String
+	let questionNumber: String
+}
