+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -2,17 +2,6 @@ import UIKit
 
 final class MovieQuizViewController: UIViewController {
 	// MARK: Все переменные
-	private var questions: [QuizQuestion] = [
-		QuizQuestion(image: "The Godfather", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-		QuizQuestion(image: "The Dark Knight", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-		QuizQuestion(image: "Kill Bill", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-		QuizQuestion(image: "The Avengers", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-		QuizQuestion(image: "Deadpool", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-		QuizQuestion(image: "The Green Knight", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-		QuizQuestion(image: "Old", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false),
-		QuizQuestion(image: "The Ice Age Adventures of Buck Wild", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false),
-		QuizQuestion(image: "Tesla", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false),
-		QuizQuestion(image: "Vivarium", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false)]
 	private var currentQuestionIndex: Int = 0
 	private var correctAnswers: Int = 0
 	private var countOfQuestions: Int = 10
@@ -103,28 +92,5 @@ final class MovieQuizViewController: UIViewController {
 		let currentQuestion = questions[currentQuestionIndex]
 		let givenAnswer = true
 		showAnswerResult(isCorrect: givenAnswer == currentQuestion.correctAnswer)
-	}
-}
-
-extension MovieQuizViewController {
-	// MARK: View модели для состояний
-	
-	/// Вью-модель для состояния "Вопрос задан"
-	private struct QuizStepViewModel {
-		let image: UIImage
-		let question: String
-		let questionNumber: String
-	}
-	/// Вью-модель для состояния "Результат квиза"
-	private struct QuizResultsViewModel {
-		let title: String
-		let text: String
-		let buttonText: String
-	}
-	/// структура вопроса
-	private struct QuizQuestion {
-		let image: String
-		let text: String
-		let correctAnswer: Bool
 	}
 }
